+++ tests/CharacterControllerTest.php
@@ -59,12 +59,25 @@ class CharacterControllerTest extends WebTestCase
         $this->assertTrue($response->headers->contains('Content-Type', 'application/json'), $response->headers);
     }
 
-    // public function testIndex()
-    // {
-    //  $this->client->request('GET', '/buildings/');
-    //     $this->assertResponseCode(200);
-    //     $this->assertJsonResponse();
-    // }
+    public function testIndex()
+    {
+         // Tests with default values
+         $this->client->request('GET', '/characters/');
+         $this->assertResponseCode(200);
+         $this->assertJsonResponse();
+         // Tests with page
+         $this->client->request('GET', '/characters/?page=1');
+         $this->assertResponseCode(200);
+         $this->assertJsonResponse();
+         // Tests with page and size
+         $this->client->request('GET', '/characters/?page=1&size=1');
+         $this->assertResponseCode(200);
+         $this->assertJsonResponse();
+         // Tests with size
+         $this->client->request('GET', '/characters/?size=1');
+         $this->assertResponseCode(200);
+         $this->assertJsonResponse();
+    }
 
     public function testBadIdentifier()
     {